refactor(sdk): organize OpenH264 test fixtures
