+++ cvat-video-openh264/tests/fixtures/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
+"""Generated and checked-in media fixtures used by the package tests."""
