+++ cvat-video-openh264/tests/test_conversion.py
@@ -8,6 +8,8 @@ import pytest
 
 import cvat_video_openh264._reader as reader
 
+from tests.helpers import convert_single_i420_pixel
+
 
 def test_copy_plane_honors_stride() -> None:
     source = ctypes.create_string_buffer(bytes((16, 17, 255, 18, 19, 255)))
@@ -15,26 +17,6 @@ def test_copy_plane_honors_stride() -> None:
     plane = reader._copy_plane(ctypes.addressof(source), width=2, height=2, stride=3)
 
     assert plane.tobytes() == bytes((16, 17, 18, 19))
-
-
-def _convert_single_i420_pixel(y: int, u: int = 128, v: int = 128) -> tuple[int, int, int]:
-    y_source = ctypes.create_string_buffer(bytes((y,)))
-    u_source = ctypes.create_string_buffer(bytes((u,)))
-    v_source = ctypes.create_string_buffer(bytes((v,)))
-    planes = (ctypes.c_void_p * 3)(
-        ctypes.addressof(y_source),
-        ctypes.addressof(u_source),
-        ctypes.addressof(v_source),
-    )
-    buffer_info = reader._BufferInfo()
-    system_buffer = buffer_info.user_data.system_buffer
-    system_buffer.width = 1
-    system_buffer.height = 1
-    system_buffer.format = 23
-    system_buffer.stride[0] = 1
-    system_buffer.stride[1] = 1
-
-    return reader._i420_to_rgb(planes, buffer_info).getpixel((0, 0))
 
 
 @pytest.mark.parametrize(
@@ -45,7 +27,7 @@ def test_i420_conversion_expands_limited_range_luma(
     limited_range_y: int,
     expected_rgb: tuple[int, int, int],
 ) -> None:
-    assert _convert_single_i420_pixel(limited_range_y) == expected_rgb
+    assert convert_single_i420_pixel(limited_range_y) == expected_rgb
 
 
 def test_i420_conversion_supports_odd_dimensions_and_owns_output() -> None:
